sdk/cortexops/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cortexops"
version = "0.1.0"
description = "Reliability infrastructure for AI agents — evaluation, observability, and regression testing"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Ashish", email = "ashishodu2023@gmail.com" },
]
keywords = [
    "llm", "agents", "evaluation", "observability",
    "langgraph", "crewai", "autogen", "ai", "testing",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
http = ["httpx>=0.27"]
llm  = ["httpx>=0.27"]
all  = ["httpx>=0.27"]
dev  = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage      = "https://cortexops.ai"
Repository    = "https://github.com/ashishodu2023/cortexops"
Documentation = "https://docs.cortexops.ai"
"Bug Tracker" = "https://github.com/ashishodu2023/cortexops/issues"
Changelog     = "https://github.com/ashishodu2023/cortexops/releases"

[project.scripts]
cortexops = "cortexops.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["cortexops"]

[tool.hatch.build.targets.sdist]
include = [
    "cortexops/",
    "tests/",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
